fuzzy_np/requirements.txt
flask>=3.0.0
python-dotenv>=0.19.2
requests>=2.31.0
flask_cors
openai>=1.8.0